refactor: move shared runtime to shared folder
